+++ crates/ghost_actor/src/ghost_protocol.rs
@@ -1,3 +1,4 @@
+use crate::*;
 use std::fmt::Debug;
 
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
@@ -90,6 +91,20 @@ pub trait GhostProtocol: 'static + Debug + Clone + Send + Sync {
             false
         }
     }
+}
+
+/// when a protocol handler received a callback for fulfilling a request
+/// it will follow this signature
+pub type GhostHandlerCb<'lt, T> = Box<dyn FnOnce(T) -> GhostResult<()> + 'lt + Send + Sync>;
+
+/// any protocol handler should implement this trait
+pub trait GhostHandler<'lt, X: 'lt + Send + Sync, P: GhostProtocol>: Send + Sync {
+    fn trigger(
+        &mut self,
+        user_data: &mut X,
+        message: P,
+        cb: Option<GhostHandlerCb<'lt, P>>,
+    ) -> GhostResult<()>;
 }
 
 #[cfg(test)]
